+++ api/src/identity-access-management/domain/usecases/reconcile-oidc-user.usecase.js
@@ -1,8 +1,8 @@
-import { AuthenticationKeyExpired } from '../../../src/identity-access-management/domain/errors.js';
-import { AuthenticationMethod } from '../../../src/identity-access-management/domain/models/AuthenticationMethod.js';
-import { MissingUserAccountError } from '../errors.js';
+import { MissingUserAccountError } from '../../../../lib/domain/errors.js';
+import { AuthenticationKeyExpired } from '../errors.js';
+import { AuthenticationMethod } from '../models/AuthenticationMethod.js';
 
-const reconcileOidcUser = async function ({
+export const reconcileOidcUser = async function ({
   authenticationKey,
   oidcAuthenticationService,
   authenticationSessionService,
@@ -48,5 +48,3 @@ const reconcileOidcUser = async function ({
 
   return { accessToken, logoutUrlUUID };
 };
-
-export { reconcileOidcUser };
