+++ api/tests/identity-access-management/unit/domain/usecases/reconcile-oidc-user.usecase.test.js
@@ -1,11 +1,11 @@
-import { MissingUserAccountError } from '../../../../lib/domain/errors.js';
-import { reconcileOidcUser } from '../../../../lib/domain/usecases/reconcile-oidc-user.js';
-import { POLE_EMPLOI } from '../../../../src/identity-access-management/domain/constants/oidc-identity-providers.js';
-import { AuthenticationKeyExpired } from '../../../../src/identity-access-management/domain/errors.js';
-import { AuthenticationMethod } from '../../../../src/identity-access-management/domain/models/AuthenticationMethod.js';
-import { catchErr, expect, sinon } from '../../../test-helper.js';
+import { MissingUserAccountError } from '../../../../../lib/domain/errors.js';
+import { POLE_EMPLOI } from '../../../../../src/identity-access-management/domain/constants/oidc-identity-providers.js';
+import { AuthenticationKeyExpired } from '../../../../../src/identity-access-management/domain/errors.js';
+import { AuthenticationMethod } from '../../../../../src/identity-access-management/domain/models/AuthenticationMethod.js';
+import { reconcileOidcUser } from '../../../../../src/identity-access-management/domain/usecases/reconcile-oidc-user.usecase.js';
+import { catchErr, expect, sinon } from '../../../../test-helper.js';
 
-describe('Unit | UseCase | reconcile-oidc-user', function () {
+describe('Unit | Identity Access Management | Domain | UseCase | reconcile-oidc-user', function () {
   context('when identityProvider is generic', function () {
     let identityProvider;
     let authenticationMethodRepository;
